requirements.txt
black==22.3.0; python_version >= '3.6'
click>=7.0.0
coverage
cryptography
flake8
funcsigs
httpretty==1.1.4
importlib-metadata
ipykernel
ipython
jupyter_client
mypy; python_version >= '3.6'
nbconvert
pyjwt
pytest
pytest-cov
pytest-mypy; python_version >= '3.5'
semver>=2.0.0,<3.0.0
setuptools_scm
six>=1.14.0
toml
twine
wheel
types-six
types-click
types-Flask
